+++ CHANGELOG.md
@@ -18,6 +18,8 @@
 * 【core   】     Base64增加encodeWithoutPadding方法（issue#I26J16@Gitee）
 * 【core   】     ExceptionUtil增加message消息包装为运行时异常的方法（pr#1253@Gitee）
 * 【core   】     DatePattern增加年月格式化常量（pr#220@Gitee）
+* 【core   】     ArrayUtil增加shuffle方法（pr#1255@Github）
+* 【core   】     ArrayUtil部分方法分离至PrimitiveArrayUtil
 
 ### Bug修复
 * 【cron   】     修复CronTimer可能死循环的问题（issue#1224@Github）
